# build/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "mongoose.Types.ObjectId": {
        "type": "string"
      },
      "AccountState": {
        "enum": ["in-used", "forbidden"],
        "type": "string"
      },
      "UserDTO": {
        "properties": {
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录创建时间"
          },
          "modifyTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录修改时间"
          },
          "deleteTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "记录删除时间"
          },
          "id": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "abc123!@#$"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "kbmjj123@gmail.com"
          },
          "role": {
            "type": "string",
            "enum": ["user", "admin"],
            "description": "用户角色，枚举类型，user代表普通用户，admin代表管理员"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true,
            "description": "刷新的token"
          },
          "accessToken": {
            "type": "string",
            "nullable": true,
            "description": "资源访问的token"
          },
          "nickName": {
            "type": "string",
            "nullable": true,
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "nullable": true,
            "description": "用户头像"
          },
          "address": {
            "allOf": [
              {
                "$ref": "#/components/schemas/mongoose.Types.ObjectId"
              }
            ],
            "nullable": true,
            "description": "用户地址id"
          },
          "loginTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "登录时间"
          },
          "logoutTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "退出时间"
          },
          "account": {
            "type": "string",
            "nullable": true,
            "description": "用户账号"
          },
          "state": {
            "$ref": "#/components/schemas/AccountState",
            "description": "用户状态"
          }
        },
        "required": ["id", "password", "email", "role"],
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_UserDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/UserDTO"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Pick_UserDTO.password-or-email_": {
        "properties": {
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "abc123!@#$"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "kbmjj123@gmail.com"
          }
        },
        "required": ["password", "email"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "UserLoginParams": {
        "$ref": "#/components/schemas/Pick_UserDTO.password-or-email_",
        "description": "用户登录所需参数"
      },
...
      "BaseObjectEntity__accessToken-string--refreshToken-string__": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "properties": {
              "refreshToken": {
                "type": "string"
              },
              "accessToken": {
                "type": "string"
              }
            },
            "required": ["refreshToken", "accessToken"],
            "type": "object"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "BaseObjectEntity_string_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_unknown_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {}
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "Pick_UserDTO.password_": {
        "properties": {
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "abc123!@#$"
          }
        },
        "required": ["password"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "UserRegisterParams": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_UserDTO.password_"
          },
          {
            "properties": {
              "token": {
                "type": "string",
                "description": "用户注册所使用的jwt token"
              }
            },
            "required": ["token"],
            "type": "object"
          }
        ]
      },
      "Pick_UserDTO.email-or-password_": {
        "properties": {
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "abc123!@#$"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "kbmjj123@gmail.com"
          }
        },
        "required": ["password", "email"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "UserQuickRegisterParams": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_UserDTO.email-or-password_"
          },
          {
            "properties": {
              "code": {
                "type": "string",
                "description": "邮箱验证码"
              }
            },
            "required": ["code"],
            "type": "object"
          }
        ],
        "description": "携带验证码进行快速注册的实体"
      },
      "Pick_UserDTO.account-or-avatar-or-nickName_": {
        "properties": {
          "nickName": {
            "type": "string",
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像"
          },
          "account": {
            "type": "string",
            "description": "用户账号"
          }
        },
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "EditUserParams": {
        "$ref": "#/components/schemas/Pick_UserDTO.account-or-avatar-or-nickName_",
        "description": "编辑用户信息需需要参数"
      },
      "BaseObjectEntity_String_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "ModifyPwdParams": {
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "新密码"
          },
          "oldPassword": {
            "type": "string",
            "description": "旧密码"
          }
        },
        "required": ["newPassword", "oldPassword"],
        "type": "object",
        "description": "修改密码参数"
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "KBMall",
    "version": "1.0.0",
    "description": "KB商城的买家端、卖家端的接口服务",
    "license": {
      "name": "ISC"
    },
    "contact": {
      "email": "kbmjj123@gmail.com",
      "name": "郑耿林"
    }
  },
  "paths": {
    "/user/login": {
      "post": {
        "operationId": "CheckUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "用户登录接口",
        "tags": ["用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginParams"
              }
            }
          }
        }
      }
    },
...
    "/user/refreshToken": {
      "patch": {
        "operationId": "RefreshToken",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity__accessToken-string--refreshToken-string__"
                }
              }
            }
          }
        },
        "description": "刷新用户的accessToken以及refreshToken，即延长用户的在线有效性",
        "tags": ["用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": ["refreshToken"],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/user/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "用户登录",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginParams"
              }
            }
          }
        }
      }
    },
    "/api/user/getRegisterLink": {
      "get": {
        "operationId": "GetRegisterLink",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_string_"
                }
              }
            }
          }
        },
        "description": "根据邮箱获取注册的链接",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/user/register": {
      "post": {
        "operationId": "RegisterNewAccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "新用户通过邮件链接进行注册",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterParams"
              }
            }
          }
        }
      }
    },
    "/api/user/resetPwd": {
      "post": {
        "operationId": "ResetPwd",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_string_"
                }
              }
            }
          }
        },
        "description": "重置密码",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  }
                },
                "required": ["password", "token"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/user/getResetPwdLink": {
      "get": {
        "operationId": "GetResetPwdLink",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_string_"
                }
              }
            }
          }
        },
        "description": "根据邮箱获取重置密码的链接",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/user/quickRegisterCode": {
      "get": {
        "operationId": "GetQuickRegisterCode",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_unknown_"
                }
              }
            }
          }
        },
        "description": "快速注册获取验证码邮件",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/user/registerWithCode": {
      "post": {
        "operationId": "RegisterWithCode",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "通过邮件验证码进行快速注册",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserQuickRegisterParams"
              }
            }
          }
        }
      }
    },
    "/api/user/edit": {
      "post": {
        "operationId": "ModifyAUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "修改用户信息，主要为用户自主修改",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditUserParams"
              }
            }
          }
        }
      }
    },
    "/api/user/modifyPwd": {
      "post": {
        "operationId": "ModifyPwd",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_String_"
                }
              }
            }
          }
        },
        "description": "根据旧密码修改新密码",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyPwdParams"
              }
            }
          }
        }
      }
    },
...
    "/api/user/refreshToken": {
      "patch": {
        "operationId": "RefreshToken",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity__accessToken-string--refreshToken-string__"
                }
              }
            }
          }
        },
        "description": "刷新用户的accessToken以及refreshToken，即延长用户的在线有效性",
        "tags": ["商城/用户模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": ["refreshToken"],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}