# dto/UserDTO.ts
import { Types } from "mongoose";
import { ISoftDeleteDTO } from "./soft-delete-dto/ISoftDeleteDTO";
import { AccountState } from "../enum/business";

export interface UserDTO extends ISoftDeleteDTO{

	id:string;
	/**
	 * 用户密码
	 * @example "abc123!@#$"
	*/
	password: string;
	/**
	 * 用户邮箱
	 * @example "kbmjj123@gmail.com"
	*/
	email: string;
	/**
	 * 用户角色，枚举类型，user代表普通用户，admin代表管理员
	 * @example user
	*/
	role: "user" | "admin";
	/**
	 * 刷新的token
	*/
	refreshToken?: string | null | undefined;
	/**
	 * 资源访问的token
	*/
	accessToken?: string | null | undefined;
	/**
	 * 用户昵称
	 * @example 昵称
	*/
	nickName?: string | null | undefined;
	/**
	 * 用户头像
	*/
	avatar?: string | null | undefined;
	/**
	 * 用户地址id
	*/
	address?: Types.ObjectId | null | undefined;
	/**
	 * 登录时间
	*/
	loginTime?: Date | null | undefined;
	/**
	 * 退出时间
	*/
	logoutTime?: Date | null | undefined;
	/**
	 * 用户账号
	 * @example 账号
	*/
	account?: string | null | undefined;
	/**
	 * 用户状态
	*/
	state?: AccountState;

	isPasswordMatched(newPwd: string): Promise<boolean>;

}
/**
 * 不带token的用户信息
*/
export type UserWithoutToken = Omit<UserDTO, 'password' | 'refreshToken' | 'accessToken' | 'logoutTime' | 'isPasswordMatched'>
/**
 * 编辑用户信息需需要参数
 */
export type EditUserParams = Pick<UserDTO, 'account' | 'avatar' | 'nickName'>

/**
 * 用户登录所需参数
 */
export type UserLoginParams = Pick<UserDTO, 'password' | 'email'>

/**
 * 修改密码参数
*/
export type ModifyPwdParams = {
	/**
	 * 旧密码
	*/
	oldPassword: string,
	/**
	 * 新密码
	*/
	newPassword: string
}

export type UserRegisterParams = Pick<UserDTO, 'password'> & {
	/**
		 * 用户注册所使用的jwt token
		 */
	token: string
}
/**
 * 携带验证码进行快速注册的实体
*/
export type UserQuickRegisterParams = Pick<UserDTO, 'email' | 'password'> & {
	/**
	 * 邮箱验证码
	*/
	code: string
}
/**
 * 变更用户状态的参数
*/
export type UserToggleEnabledParams = Pick<UserDTO, 'state'>